+++ src/main/java/br/com/fta/transaction/TransactionRepository.java
@@ -1,4 +1,4 @@
-package br.com.fta.repository;
+package br.com.fta.transaction;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -6,8 +6,6 @@ import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-
-import br.com.fta.model.Transaction;
 
 @Repository
 public interface TransactionRepository extends MongoRepository<Transaction, String>{
